osd/scrub: create a separate chunk size conf for shallow scrubs
Using the existing common default chunk size for scrubs that are
not deep scrubs is wasteful: a high ratio of inter-OSD messages
per chunk, while the actual OSD work per chunk is minimal.

+++ src/osd/scrubber/pg_scrubber.h
@@ -941,7 +941,10 @@ class PgScrubber : public ScrubPgIF,
 
     [[nodiscard]] bool was_preempted() const { return m_preempted; }
 
-    [[nodiscard]] size_t chunk_divisor() const { return m_size_divisor; }
+    [[nodiscard]] int chunk_divisor() const
+    {
+      return static_cast<int>(m_size_divisor);
+    }
 
     void reset();
